+++ cpp/ampls/src/ampls.cpp
@@ -110,21 +110,16 @@ std::string impl::Records::getRecordedEntities(bool exportToAMPL) {
     for (int i = 0; i < values.size(); i++)
       vars_[i].value(values[i]);
   }
-
-  
-  if (exportToAMPL) {
-    for (auto e : entities_)
-    {
-      if (!e->exportedToAMPL_)
-      {
-        ss << e->toAMPLString(vmap, cmap, *this) << std::endl;
+  for (auto& entity : entities_)
+  {
+    ampls::Entity* e;
+    if (entity.first)
+      e = &vars_[entity.second];
+    else
+      e= &cons_[entity.second];
+    if (exportToAMPL && (!e->exportedToAMPL_))
         e->exportedToAMPL_ = true;
-      }
-    }
-  }
-  else {
-    for (auto e : entities_)
-     ss << e->toAMPLString(vmap, cmap, *this) << std::endl;
+    ss << e->toAMPLString(vmap, cmap, *this) << std::endl;
   }
   return ss.str();
 }
@@ -293,7 +288,7 @@ namespace impl {
     int status = doAddCut(c, type);
     if (status) 
       throw ampls::AMPLSolverException::format("Error while adding cut!");
-    c.solverIndex(getValue(ampls::Value::CBValue::N_ROWS).integer - 1);
+    c.solverIndex(getValueImpl(ampls::Value::CBValue::N_ROWS).integer - 1);
     return c;
   }